packages/cli/templates/fastapi/requirements.txt
auth-kit-fastapi>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
httpx>=0.26.0
email-validator>=2.1.0